refactor(transport): move transport consumer
