+++ app/consumer/app.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/alhamsya/boilerplate-go/app"
 	"github.com/alhamsya/boilerplate-go/lib/helpers/custom_log"
-	"github.com/alhamsya/boilerplate-go/transport/inter/consumer/handler"
+	"github.com/alhamsya/boilerplate-go/transport/consumer/handler"
 )
 
 func main() {
