+++ app/init.go
@@ -11,9 +11,9 @@ import (
 	"github.com/alhamsya/boilerplate-go/lib/helpers/database"
 	"github.com/alhamsya/boilerplate-go/lib/utils"
 	"github.com/alhamsya/boilerplate-go/middleware/rest"
+	"github.com/alhamsya/boilerplate-go/transport/consumer/routers"
 	"github.com/alhamsya/boilerplate-go/transport/cron/routers"
 	"github.com/alhamsya/boilerplate-go/transport/grpc/routers"
-	"github.com/alhamsya/boilerplate-go/transport/inter/consumer/routers"
 	"github.com/alhamsya/boilerplate-go/transport/rest/routers"
 	"github.com/alhamsya/boilerplate-go/usecase/consumer"
 	"github.com/alhamsya/boilerplate-go/usecase/cron"
