+++ transport/consumer/handler/types.go
@@ -3,7 +3,7 @@ package consumerHandler
 import (
 	"cloud.google.com/go/pubsub"
 	"github.com/alhamsya/boilerplate-go/lib/helpers/config"
-	"github.com/alhamsya/boilerplate-go/transport/inter/consumer/routers"
+	"github.com/alhamsya/boilerplate-go/transport/consumer/routers"
 )
 
 type Handler struct {
